+++ src/Settings/PreferencesPageTemplate.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Ashar Khan <ashar786khan@gmail.com>
+ * Copyright (C) 2019-2021 Ashar Khan <ashar786khan@gmail.com>
  *
  * This file is part of CP Editor.
  *
@@ -27,9 +27,9 @@ class PreferencesPageTemplate : public PreferencesGridPage
   public:
     explicit PreferencesPageTemplate(QStringList opts, bool alignTop = true, QWidget *parent = nullptr);
 
-    virtual QStringList content() override;
+    QStringList content() override;
 
-    virtual void setPath(const QString &path, const QString &trPath) override;
+    void setPath(const QString &path, const QString &trPath) override;
 
   private:
     bool areSettingsChanged() override;
@@ -39,7 +39,7 @@ class PreferencesPageTemplate : public PreferencesGridPage
 
     void onDependencyUpdated(const QString &settingName);
 
-  protected:
+  private:
     QStringList options;
     QVector<ValueWidget *> widgets;
 };