+++ src/Settings/SettingsManager.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Ashar Khan <ashar786khan@gmail.com>
+ * Copyright (C) 2019-2021 Ashar Khan <ashar786khan@gmail.com>
  *
  * This file is part of CP Editor.
  *
@@ -122,7 +122,7 @@ void SettingsManager::deinit()
 {
     set("Total Usage Time", get("Total Usage Time").toInt() + QDateTime::currentSecsSinceEpoch() - startTime);
 
-    saveSettings(Util::configFilePath(configFileLocations[0]));
+    saveSettings(QString());
 
     delete cur;
     delete def;
@@ -161,9 +161,11 @@ void SettingsManager::loadSettings(const QString &path)
 
 void SettingsManager::saveSettings(const QString &path)
 {
-    LOG_INFO("Start saving settings to " + path);
+    const auto savePath = path.isEmpty() ? Util::configFilePath(configFileLocations[0]) : path;
 
-    QSettings setting(path, QSettings::IniFormat);
+    LOG_INFO("Start saving settings to " + savePath);
+
+    QSettings setting(savePath, QSettings::IniFormat);
     setting.clear(); // Otherwise SettingsManager::remove won't work
     save(setting, "", SettingsInfo::getSettings());
 
@@ -172,7 +174,7 @@ void SettingsManager::saveSettings(const QString &path)
 
     setting.sync();
 
-    LOG_INFO("Settings have been saved to " + path);
+    LOG_INFO("Settings have been saved to " + savePath);
 }
 
 QVariant SettingsManager::get(QString const &key, bool alwaysDefault)