+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 -   Now you can toggle fullscreen mode by pressing F11 or clicking View->Full Screen. (#642, #660 and #670)
 -   In test cases, you can copy the output to the expected. (#688)
 -   Copy test cases between different tabs. (#688)
+-   Now settings are saved to disk as soon as new settings are applied. (#590 and #722)
+-   x86 release for 32-bit Windows. (#719 and #723)
 
 ### Fixed